Try to fix segfault with high requets count
It looks like its coming from the fact that TSRM reuse ressources for the same thread.
So even if two goroutines handle two different requests it can end up on the same thread reusing the same sapi globals.
But then we kind of but not really clean them so it can segfault because the heap is corrupted.

Cleaning them (with `ts_free_thread`) at the end of each requests causes segfault as well but cleaning the context only every X requests works?